requirements.txt
streamlit==1.51.0
python-dotenv==1.2.1
langchain>=0.0.350
openai>=1.10.0,<2.0.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5

# If you need additional pinned dependencies, add them below.
